Modified all instances of Duke/duke back to TrashBot/trashbot; A-Jar Done
