--- src/main/java/org/duke/exceptions/DukeException.java
@@ -1,7 +1,0 @@
-package org.duke.exceptions;
-
-public class DukeException extends Exception {
-    public DukeException(String message) {
-        super(message);
-    }
-}
