--- src/main/java/org/duke/exceptions/EmptyDescriptionException.java
@@ -1,7 +1,0 @@
-package org.duke.exceptions;
-
-public class EmptyDescriptionException extends DukeException {
-    public EmptyDescriptionException(String task) {
-        super("The task " + task + " can't be empty!");
-    }
-}
