--- src/main/java/org/duke/exceptions/InvalidFormatException.java
@@ -1,7 +1,0 @@
-package org.duke.exceptions;
-
-public class InvalidFormatException extends DukeException {
-    public InvalidFormatException(String message) {
-        super(message);
-    }
-}
