--- src/main/java/org/duke/exceptions/StorageException.java
@@ -1,7 +1,0 @@
-package org.duke.exceptions;
-
-public class StorageException extends DukeException {
-    public StorageException(String message) {
-        super(message);
-    }
-}
