--- src/main/java/org/duke/tasks/Todo.java
@@ -1,12 +1,0 @@
-package org.duke.tasks;
-
-public class Todo extends Task {
-    public Todo(String input) {
-        super(input.substring(5)); // get rid of "todo "
-    }
-
-    @Override
-    public String toString() {
-        return "[T]" + super.toString();
-    }
-}
