+++ src/main/java/org/trashbot/exceptions/DukeException.java
@@ -1,0 +1,7 @@
+package org.trashbot.exceptions;
+
+public class DukeException extends Exception {
+    public DukeException(String message) {
+        super(message);
+    }
+}
