+++ src/main/java/org/trashbot/exceptions/EmptyDescriptionException.java
@@ -1,0 +1,7 @@
+package org.trashbot.exceptions;
+
+public class EmptyDescriptionException extends DukeException {
+    public EmptyDescriptionException(String task) {
+        super("The task " + task + " can't be empty!");
+    }
+}
