+++ src/main/java/org/trashbot/exceptions/InvalidFormatException.java
@@ -1,0 +1,7 @@
+package org.trashbot.exceptions;
+
+public class InvalidFormatException extends DukeException {
+    public InvalidFormatException(String message) {
+        super(message);
+    }
+}
