+++ src/main/java/org/trashbot/exceptions/StorageException.java
@@ -1,0 +1,7 @@
+package org.trashbot.exceptions;
+
+public class StorageException extends DukeException {
+    public StorageException(String message) {
+        super(message);
+    }
+}
