+++ src/main/java/org/trashbot/exceptions/UnknownInputException.java
@@ -1,0 +1,7 @@
+package org.trashbot.exceptions;
+
+public class UnknownInputException extends DukeException {
+    public UnknownInputException(String input) {
+        super("The input " + input + " is unknown!\n" + " Available input: todo, deadline, event, mark, unmark, list, bye");
+    }
+}
