+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
 ____________________________________________________________
- Hello! I'm core.org.duke.TrashBot
+ Hello! I'm core.org.trashbot.TrashBot
  What can I do you for?
 ____________________________________________________________
 ____________________________________________________________
